Add visualization tool for filters and fix test
Former-commit-id: 8f5cb20329deb17a1952c05cd8218d8d91be9bf8

+++ code/test/test_metrics/test_mask_metric.py
@@ -11,9 +11,9 @@ def test_mask_metric(dataset):
 
     dataloader = DataLoader(dataset, batch_size=4)
     batch = next(iter(dataloader))
-    perfect_out = torch.nn.functional.one_hot(batch["target"])[
-        ..., :133
-    ]
+    perfect_out = torch.nn.functional.one_hot(
+        batch["target"].to(dtype=torch.long)
+    )[..., :133]
     # Model output shape is [B, C, H, W]
     perfect_out = perfect_out.permute(0, 3, 1, 2)
